+++ docs/history.rst
@@ -3,6 +3,7 @@ History
 
 Latest
 ------
+- REF: Store `grid_mapping` in `encoding` instead of `attrs` (issue #282)
 
 0.3.2
 -----